libft_commit

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: youjeon <youjeon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/10 11:30:45 by youjeon           #+#    #+#             */
-/*   Updated: 2021/05/10 13:10:09 by youjeon          ###   ########.fr       */
+/*   Updated: 2021/05/13 18:38:09 by youjeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_bzero(void *str, size_t n)
 {
 	unsigned char	*ptr_str;
-	int				i;
+	size_t			i;
 
 	ptr_str = str;
 	i = 0;
