+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: youjeon <youjeon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/07 15:44:55 by youjeon           #+#    #+#             */
-/*   Updated: 2021/05/10 11:30:17 by youjeon          ###   ########.fr       */
+/*   Updated: 2021/05/13 18:38:58 by youjeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memset(void *dest, int c, size_t len)
 {
 	unsigned char	*ptr_dest;
 	unsigned char	trans_c;
-	int				i;
+	size_t			i;
 
 	ptr_dest = dest;
 	trans_c = c;
